@@ -1,17 +1,18 @@
 import traceback
 from abc import ABC, abstractmethod
+from collections.abc import Callable
 from time import time
-from typing import Any, Callable, Generic, Optional, Union
+from typing import Any, Generic
 
 from qcio import (
     CalcType,
-    FileInput,
+    Data,
+    DataType,
     Files,
-    InputType,
-    ProgramOutput,
+    FileSpec,
     Results,
-    ResultsType,
-    StructuredInputs,
+    SpecType,
+    StructuredSpecs,
 )
 from qcio.helper_types import StrOrPath
 
@@ -28,7 +29,7 @@ __all__ = ["BaseAdapter", "registry"]
 registry = {}
 
 
-class BaseAdapter(ABC, Generic[InputType, ResultsType]):
+class BaseAdapter(ABC, Generic[SpecType, DataType]):
     """Base class for all adapters."""
 
     # Whether to this program reads or writes files to disk.
@@ -40,7 +41,7 @@ class BaseAdapter(ABC, Generic[InputType, ResultsType]):
     uses_files = True
     program: str  # All subclasses must define this attribute.
 
-    def program_version(self, stdout: Optional[str]) -> Optional[str]:
+    def program_version(self, stdout: str | None) -> str | None:
         """Return program version. Adapters should override this method.
 
         Args:
@@ -53,49 +54,49 @@ class BaseAdapter(ABC, Generic[InputType, ResultsType]):
         return None
 
     @abstractmethod
-    def validate_input(self, input_data: InputType) -> None:
+    def validate_input(self, input_data: SpecType) -> None:
         """Validate input_data to ensure compatibility with adapter.
         Adapters should override this method.
         """
         raise NotImplementedError
 
     @abstractmethod
-    def compute_results(
+    def compute_data(
         self,
-        input_data: InputType,
-        update_func: Optional[Callable] = None,
-        update_interval: Optional[float] = None,
+        input_data: SpecType,
+        update_func: Callable | None = None,
+        update_interval: float | None = None,
         **kwargs,
-    ) -> tuple[ResultsType, str]:
+    ) -> tuple[DataType, str]:
         """Subclasses should implement this method with custom compute logic."""
         raise NotImplementedError
 
     def compute(
         self,
-        input_data: InputType,
+        input_data: SpecType,
         *,
-        scratch_dir: Optional[StrOrPath] = None,
+        scratch_dir: StrOrPath | None = None,
         rm_scratch_dir: bool = True,
-        collect_stdout: bool = True,
+        collect_logs: bool = True,
         collect_files: bool = False,
         collect_wfn: bool = False,
-        update_func: Optional[Callable] = None,
-        update_interval: Optional[float] = None,
-        print_stdout: bool = False,
+        update_func: Callable | None = None,
+        update_interval: float | None = None,
+        print_logs: bool = False,
         raise_exc: bool = True,
         propagate_wfn: bool = False,
         **adapter_kwargs,
-    ) -> ProgramOutput[InputType, ResultsType]:
+    ) -> Results[SpecType, DataType]:
         """Compute the given input using the adapter's program.
 
         Args:
-            input_data: A qcio input object for a computation. E.g. A FileInput,
-                ProgramInput or DualProgramInput.
+            input_data: A qcio input object for a computation. E.g. A FileSpec,
+                CalcSpec or CompositeCalcSpec.
             scratch_dir: The scratch directory for the program. If None, a new directory
                 is created in the system default temporary directory. If rm_scratch_dir
                 is True this directory will be deleted after the program finishes.
             rm_scratch_dir: Delete the scratch directory after the program exits.
-            collect_stdout: Whether to collect stdout/stderr from the program as output.
+            collect_logs: Whether to collect stdout/stderr from the program as output.
                 Failed computations will always collect stdout/stderr.
             collect_files: Collect all files generated by the QC program as output.
             collect_wfn: Collect the wavefunction file(s) from the calculation.
@@ -106,9 +107,9 @@ class BaseAdapter(ABC, Generic[InputType, ResultsType]):
                 argument.
             update_interval: The minimum time in seconds between calls to the
                 update_func.
-            print_stdout: Whether to print stdout/stderr to the terminal in real time as
+            print_logs: Whether to print stdout/stderr to the terminal in real time as
                 the program executes. Will be ignored if an update_func passed.
-            raise_exc: If False, qcop will return a ProgramOutput object when the QC
+            raise_exc: If False, qcop will return a Results object when the QC
                 program fails rather than raise an exception.
             propagate_wfn: For any adapter performing a sequential task, such
                 as a geometry optimization, propagate the wavefunction from the previous
@@ -120,7 +121,7 @@ class BaseAdapter(ABC, Generic[InputType, ResultsType]):
                 qcng.compute().
 
         Returns:
-            A ProgramOutput object containing the results of the computation.
+            A Results object containing the results of the computation.
 
         Raises:
             AdapterNotFoundError: If the program is not supported (i.e., no Adapter
@@ -133,7 +134,7 @@ class BaseAdapter(ABC, Generic[InputType, ResultsType]):
             QCEngineError: If QCEngine performs the computation raises an error.
         """
         # Print stdout to terminal in real time as program executes
-        if print_stdout and update_func is None:
+        if print_logs and update_func is None:
             update_func, update_interval = (
                 lambda _, stdout_new: print(stdout_new),
                 0.1,
@@ -145,51 +146,49 @@ class BaseAdapter(ABC, Generic[InputType, ResultsType]):
                 input_data.save_files()
 
             # Define outputs
-            output_dict: dict[str, Any] = {}
-            stdout: Optional[str] = None
-            results: Results
-            exc: Optional[QCOPBaseError] = None
-            program_version: Optional[str] = None
+            results_dict: dict[str, Any] = {}
+            logs: str | None = None
+            data: Data
+            exc: QCOPBaseError | None = None
+            program_version: str | None = None
 
             start = time()
             try:
                 # Validate input object
                 self.validate_input(input_data)
 
-                # Execute the program. results will be None if FileInput
-                results, stdout = self.compute_results(
+                # Execute the program. results will be None if FileSpec
+                data, logs = self.compute_data(
                     input_data,
                     update_func,
                     update_interval,
                     propagate_wfn=propagate_wfn,
                     **adapter_kwargs,
                 )
-                # None value covers FileInput case
-                # TODO: Is there a type safe way to handle this??
-                output_dict["success"] = True
+                # None value covers FileSpec case
+                results_dict["success"] = True
 
                 # Optionally collect wavefunction file
                 if collect_wfn and not collect_files:
-                    output_dict["files"] = self.collect_wfn()
+                    results_dict["files"] = self.collect_wfn()
 
             except QCOPBaseError as e:
-                # TODO: Is there a type safe way to handle this??
                 exc = e
-                output_dict["success"] = False
-                # Any half-completed results
-                results = getattr(e, "results") or Files()
-                stdout = getattr(e, "stdout", stdout)
-                # For mypy because e.stdout is not of a known type
-                stdout = str(stdout) if stdout is not None else None
-                output_dict["traceback"] = traceback.format_exc()
+                results_dict["success"] = False
+                # Any half-completed data
+                data = getattr(e, "data") or Files()
+                logs = getattr(e, "logs", logs) or logs
+                # For mypy because e.logs is not of a known type
+                logs = str(logs) if logs is not None else None
+                results_dict["traceback"] = traceback.format_exc()
 
             wall_time = time() - start
 
             # Check for parsed version in extras at default location
-            program_version = results.extras.get("program_version")
+            program_version = data.extras.get("program_version")
             if not program_version:
                 try:
-                    program_version = self.program_version(stdout)
+                    program_version = self.program_version(logs)
                 except ProgramNotFoundError:
                     pass  # program_version = None set above
 
@@ -201,37 +200,37 @@ class BaseAdapter(ABC, Generic[InputType, ResultsType]):
                 wall_time,
             )
 
-            # Always collect for failures; otherwise obey collect_stdout
-            stdout = stdout if not output_dict["success"] or collect_stdout else None
+            # Always collect for failures; otherwise obey collect_logs
+            logs = logs if not results_dict["success"] or collect_logs else None
 
             # Construct output object
-            output_dict.update(
+            results_dict.update(
                 {
                     "input_data": input_data,
-                    "stdout": stdout,
-                    "results": results,
+                    "logs": logs,
+                    "data": data,
                     "provenance": provenance,
                 }
             )
-            output_obj = ProgramOutput[InputType, ResultsType](**output_dict)
+            results = Results[SpecType, DataType](**results_dict)
 
             # Collect files generated by the program
-            if self.uses_files and (collect_files or type(input_data) is FileInput):
-                output_obj.results.add_files(
+            if self.uses_files and (collect_files or type(input_data) is FileSpec):
+                results.data.add_files(
                     final_scratch_dir,
                     recursive=True,
                     exclude=list(input_data.files.keys()),
                 )
 
-        # Append ProgramOutput to exception and raise if raise_exc=True
+        # Append Results to exception and raise if raise_exc=True
         # Helpful for BigChem and ChemCloud exception handling
         if raise_exc and exc:
-            exc.program_output = output_obj
+            exc.results = results
             raise exc
 
-        return output_obj
+        return results
 
-    def collect_wfn(self) -> dict[str, Union[str, bytes]]:
+    def collect_wfn(self) -> dict[str, str | bytes]:
         """Collect the wavefunction file(s) from the scratch_dir.
 
         Returns:
@@ -245,7 +244,7 @@ class BaseAdapter(ABC, Generic[InputType, ResultsType]):
         )
 
 
-class ProgramAdapter(BaseAdapter, Generic[InputType, ResultsType]):
+class ProgramAdapter(BaseAdapter, Generic[SpecType, DataType]):
     """Base adapter for all program adapters (all but FileAdaptor)."""
 
     supported_calctypes: list[
@@ -272,7 +271,7 @@ class ProgramAdapter(BaseAdapter, Generic[InputType, ResultsType]):
             )
 
     @abstractmethod
-    def program_version(self, stdout: Optional[str]) -> str:
+    def program_version(self, stdout: str | None) -> str:
         """Get the version of the program.
 
         Args:
@@ -284,17 +283,17 @@ class ProgramAdapter(BaseAdapter, Generic[InputType, ResultsType]):
         """
 
     @abstractmethod
-    def compute_results(
+    def compute_data(
         self,
-        input_data: InputType,
-        update_func: Optional[Callable] = None,
-        update_interval: Optional[float] = None,
+        input_data: SpecType,
+        update_func: Callable | None = None,
+        update_interval: float | None = None,
         **kwargs,
-    ) -> tuple[ResultsType, str]:
-        """All ProgramAdapters must return a ResultsType."""
+    ) -> tuple[DataType, str]:
+        """All ProgramAdapters must return a DataType."""
         raise NotImplementedError
 
-    def validate_input(self, input_data: StructuredInputs) -> None:
+    def validate_input(self, input_data: StructuredSpecs) -> None:
         """Validate the input object for compatibility with the adapter.
 
         Args:
