@@ -1,14 +1,14 @@
 import importlib
-from typing import Callable, Optional
+from collections.abc import Callable
 
-from qcio import CalcType, ProgramInput, SinglePointResults
+from qcio import CalcSpec, CalcType, SinglePointData
 
 from qcop.exceptions import ExternalProgramError, ProgramNotFoundError
 
 from .base import ProgramAdapter
 
 
-class TeraChemFEAdapter(ProgramAdapter[ProgramInput, SinglePointResults]):
+class TeraChemFEAdapter(ProgramAdapter[CalcSpec, SinglePointData]):
     """Adapter for TeraChem's Protocol Buffer Server and Frontend file server."""
 
     supported_calctypes = [CalcType.energy, CalcType.gradient]
@@ -35,27 +35,27 @@ class TeraChemFEAdapter(ProgramAdapter[ProgramInput, SinglePointResults]):
                 ),
             )
 
-    def program_version(self, stdout: Optional[str] = None) -> str:
+    def program_version(self, stdout: str | None = None) -> str:
         """Program version is not available via the PB server."""
         return ""
 
-    def compute_results(
+    def compute_data(
         self,
-        input_data: ProgramInput,
-        update_func: Optional[Callable] = None,
-        update_interval: Optional[float] = None,
+        input_data: CalcSpec,
+        update_func: Callable | None = None,
+        update_interval: float | None = None,
         **kwargs,
-    ) -> tuple[SinglePointResults, str]:
+    ) -> tuple[SinglePointData, str]:
         """Execute TeraChem on the given input.
 
         Args:
-            input_data: The qcio ProgramInput object for a computation.
+            input_data: The qcio CalcSpec object for a computation.
             update_func: A callback function to call as the program executes.
             update_interval: The minimum time in seconds between calls to the
                 update_func.
 
         Returns:
-            A tuple of SinglePointResults and the stdout str.
+            A tuple of SinglePointData and the stdout str.
         """
         try:
             with self.client() as client:
@@ -63,9 +63,9 @@ class TeraChemFEAdapter(ProgramAdapter[ProgramInput, SinglePointResults]):
         except self.tcpb.exceptions.TCPBError as e:
             exc = ExternalProgramError(
                 program=self.program,
-                # Pass stdout to .compute() via the exception
+                # Pass logs to .compute() via the exception
                 # Will only exist for TeraChemFrontendAdapter
-                stdout=e.program_output.stdout,
+                logs=e.results.logs,
             )
 
             raise exc
@@ -73,6 +73,6 @@ class TeraChemFEAdapter(ProgramAdapter[ProgramInput, SinglePointResults]):
         else:
             # Write files to disk to be collected by BaseAdapter.compute()
             # Used only for TeraChemFrontendAdapter
-            prog_output.results.save_files()
+            prog_output.data.save_files()
 
-        return prog_output.results, prog_output.stdout
+        return prog_output.data, prog_output.logs
