@@ -1,17 +1,16 @@
 """Experimental exception hierarchy. This may be too complex and unhelpful for now"""
 
-from typing import Optional
 
-from qcio import ProgramOutput, Results
+from qcio import Data, Results
 
 
 class QCOPBaseError(Exception):
     """
     Base class for all qcop exceptions.
 
-    All QCOP exceptions must eventually have a non-None program_output attribute.
-    Lower-level code may leave program_output as None; the top-level compute() method
-    should attach the final ProgramOutput before propagating the error. If some results
+    All QCOP exceptions must eventually have a non-None results attribute.
+    Lower-level code may leave results as None; the top-level compute() method
+    should attach the final Results before propagating the error. If some results
     were computed before the error occurred, they should be attached to the exception
     as well.
     """
@@ -19,14 +18,14 @@ class QCOPBaseError(Exception):
     def __init__(
         self,
         message: str,
-        program_output: Optional[ProgramOutput] = None,
-        results: Optional[Results] = None,
+        results: Results | None = None,
+        data: Data | None = None,
     ):
         # Pass everything as positional arguments so they are captured in .args
         # Required for pickling and other serialization methods including celery.
-        super().__init__(message, program_output, results)
-        self.program_output = program_output
+        super().__init__(message, results, data)
         self.results = results
+        self.data = data
 
     def __str__(self):
         # Only the message is shown in the string representation.
@@ -35,7 +34,7 @@ class QCOPBaseError(Exception):
     def __repr__(self):
         return (
             f"{self.__class__.__name__}(message={self.args[0]!r}, "
-            f"program_output={self.program_output!r}, results={self.results!r})"
+            f"results={self.results!r}, data={self.data!r})"
         )
 
 
@@ -54,12 +53,12 @@ class AdapterNotFoundError(AdapterError):
     def __init__(
         self,
         program: str,
-        message: Optional[str] = None,
-        program_output: Optional[ProgramOutput] = None,
+        message: str | None = None,
+        results: Results | None = None,
     ):
         if message is None:
             message = f"No adapter found for program '{program}'."
-        super().__init__(message, program_output, None)
+        super().__init__(message, results, None)
         self.program = program
 
     def __repr__(self):
@@ -72,12 +71,12 @@ class AdapterInputError(AdapterError):
     def __init__(
         self,
         program: str,
-        message: Optional[str] = None,
-        program_output: Optional[ProgramOutput] = None,
+        message: str | None = None,
+        results: Results | None = None,
     ):
         if message is None:
             message = f"Invalid inputs for program '{program}'."
-        super().__init__(message, program_output, None)
+        super().__init__(message, results, None)
         self.program = program
 
     def __repr__(self):
@@ -98,23 +97,23 @@ class ExternalProgramError(QCOPBaseError):
     def __init__(
         self,
         program: str,
-        message: Optional[str] = None,
-        program_output: Optional[ProgramOutput] = None,
-        results: Optional[Results] = None,
-        original_exception: Optional[Exception] = None,
-        stdout: Optional[str] = None,
+        message: str | None = None,
+        results: Results | None = None,
+        data: Data | None = None,
+        original_exception: Exception | None = None,
+        logs: str | None = None,
     ):
         if message is None:
             message = f"External program '{program}' failed."
-        super().__init__(message, program_output, results)
+        super().__init__(message, results, data)
         self.program = program
         self.original_exception = original_exception
-        self.stdout = stdout
+        self.logs = logs
 
     def __repr__(self):
         return (
             f"{super().__repr__()}, program={self.program!r}, "
-            f"original_exception={self.original_exception!r}, stdout={self.stdout!r}"
+            f"original_exception={self.original_exception!r}, logs={self.logs!r}"
         )
 
 
@@ -124,17 +123,17 @@ class ProgramNotFoundError(ExternalProgramError):
     def __init__(
         self,
         program: str,
-        message: Optional[str] = None,
-        program_output: Optional[ProgramOutput] = None,
-        install_msg: Optional[str] = None,
+        message: str | None = None,
+        results: Results | None = None,
+        install_msg: str | None = None,
     ):
         if message is None:
             message = (
                 install_msg
                 or f"Program not found: '{program}'. Please install it and ensure it is on your PATH."
             )
-        # Call ExternalProgramError with results, original_exception, and stdout defaulting to None.
-        super().__init__(program, message, program_output, None, None, None)
+        # Call ExternalProgramError with results, original_exception, and logs defaulting to None.
+        super().__init__(program, message, results, None, None, None)
         self.program = program
         self.install_msg = install_msg
 
