@@ -1,15 +1,17 @@
 """Top level compute functions for qcop."""
 
 import traceback
-from typing import Any, Callable, Optional, Union
+from collections.abc import Callable
+from typing import Any
+from warnings import warn
 
 from qcio import (
+    CalcSpec,
     CalcType,
     Files,
-    InputType,
     Model,
-    ProgramInput,
-    ProgramOutput,
+    Results,
+    SpecType,
     Structure,
 )
 from qcio.helper_types import StrOrPath
@@ -22,21 +24,21 @@ from .utils import get_adapter, inherit_docstring_from
 @inherit_docstring_from(BaseAdapter.compute)
 def compute(
     program: str,
-    input_data: InputType,
+    input_data: SpecType,
     *,
-    scratch_dir: Optional[StrOrPath] = None,
+    scratch_dir: StrOrPath | None = None,
     rm_scratch_dir: bool = True,
-    collect_stdout: bool = True,
+    collect_logs: bool = True,
     collect_files: bool = False,
     collect_wfn: bool = False,
-    update_func: Optional[Callable] = None,
-    update_interval: Optional[float] = None,
-    print_stdout: bool = False,
+    update_func: Callable | None = None,
+    update_interval: float | None = None,
+    print_logs: bool = False,
     raise_exc: bool = True,
     propagate_wfn: bool = False,
     qcng_fallback: bool = True,
     **adapter_kwargs,
-) -> ProgramOutput:
+) -> Results:
     """Use the given program to compute on the given input.
 
     See BaseAdapter.compute for more details.
@@ -44,29 +46,44 @@ def compute(
     try:
         adapter = get_adapter(program, input_data, qcng_fallback)
     except (AdapterNotFoundError, ProgramNotFoundError) as e:
-        # Add program_output to the exception
-        output_obj = ProgramOutput[type(input_data), Files](  # type: ignore
+        # Add results to the exception
+        output_obj = Results[type(input_data), Files](  # type: ignore
             input_data=input_data,
-            results=Files(),
+            data=Files(),
             success=False,
             provenance={"program": program},
             traceback=traceback.format_exc(),
         )
-        e.program_output = output_obj
+        e.results = output_obj
         e.args = (*e.args, output_obj)
         raise e
 
     else:
+        if "collect_stdout" in adapter_kwargs:
+            warn(
+                "`collect_stdout` is deprecated; use `collect_logs` instead.",
+                DeprecationWarning,
+                stacklevel=2,
+            )
+            collect_logs = adapter_kwargs.pop("collect_stdout")
+        if "print_stdout" in adapter_kwargs:
+            warn(
+                "`print_stdout` is deprecated; use `print_logs` instead.",
+                DeprecationWarning,
+                stacklevel=2,
+            )
+            print_logs = adapter_kwargs.pop("print_stdout")
+
         return adapter.compute(
             input_data,
             scratch_dir=scratch_dir,
             rm_scratch_dir=rm_scratch_dir,
-            collect_stdout=collect_stdout,
+            collect_logs=collect_logs,
             collect_files=collect_files,
             collect_wfn=collect_wfn,
             update_func=update_func,
             update_interval=update_interval,
-            print_stdout=print_stdout,
+            print_logs=print_logs,
             raise_exc=raise_exc,
             propagate_wfn=propagate_wfn,
             **adapter_kwargs,
@@ -77,14 +94,14 @@ def compute_args(
     program: str,
     structure: Structure,
     *,
-    calctype: Union[str, CalcType],
-    model: Union[dict[str, str], Model],
-    keywords: Optional[dict[str, Any]] = None,
-    files: Optional[Union[dict[str, Union[str, bytes]], Files]] = None,
-    extras: Optional[dict[str, Any]] = None,
+    calctype: str | CalcType,
+    model: dict[str, str] | Model,
+    keywords: dict[str, Any] | None = None,
+    files: dict[str, str | bytes] | Files | None = None,
+    extras: dict[str, Any] | None = None,
     **kwargs,
-) -> ProgramOutput:
-    """Compute function that accepts independent argument for a ProgramInput.
+) -> Results:
+    """Compute function that accepts independent argument for a CalcSpec.
 
     Args:
         program: The program to run.
@@ -106,7 +123,7 @@ def compute_args(
     if isinstance(files, Files):  # Check in case Files object is passed instead of dict
         files = files.files
 
-    input_data = ProgramInput(
+    input_data = CalcSpec(
         calctype=calctype,  # type: ignore
         structure=structure,
         model=model,  # type: ignore
