@@ -1,8 +1,8 @@
-"""An example of how to create a FileInput object for a QC Program."""
+"""An example of how to create a FileSpec object for a QC Program."""
 
 from pathlib import Path
 
-from qcio import FileInput, Structure
+from qcio import FileSpec, Structure
 
 from qcop import compute
 
@@ -13,19 +13,19 @@ inp_file = Path("path/to/tc.in").read_text()  # Or your own function to create t
 structure = Structure.open("path/to/my/mol.xyz")
 xyz_str = structure.to_xyz()  # type: ignore
 
-# Create a FileInput object for TeraChem
-file_inp = FileInput(
+# Create a FileSpec object for TeraChem
+spec = FileSpec(
     files={"tc.in": inp_file, "coords.xyz": xyz_str}, cmdline_args=["tc.in"]
 )
 
 # This will write the files to disk in a temporary directory and then run
 # "terachem tc.in" in that directory.
-output = compute("terachem", file_inp, print_stdout=True)
+results = compute("terachem", spec, print_logs=True)
 
 # Data
-output.stdout
-output.input_data
-output.results.files  # Has all the files terachem creates
-output.results.files.keys()  # Print out file names
+results.logs
+results.input_data
+results.data.files  # Has all the files terachem creates
+results.data.files.keys()  # Print out file names
 # Saves all outputs with the exact structure produced by the QC program
-output.results.save_files("to/this/directory")
+results.data.save_files("to/this/directory")
