@@ -1,4 +1,4 @@
-from qcio import CalcType, ProgramInput, Structure
+from qcio import CalcSpec, CalcType, Structure
 
 from qcop import compute, exceptions
 
@@ -13,8 +13,8 @@ structure = Structure(
         [1.14668581, -0.45032174, -1.35474466],
     ],
 )
-# Define the program input
-pi = ProgramInput(
+# Define the calcspec
+spec = CalcSpec(
     structure=structure,
     calctype=CalcType.gradient,
     model={"method": "hf", "basis": "sto-3g"},  # type: ignore
@@ -23,22 +23,22 @@ pi = ProgramInput(
 
 # Run the calculation
 try:
-    # po is instance of ProgramOutput
-    po = compute("terachem", pi, collect_files=True)
+    # results is instance of Results
+    results = compute("terachem", spec, collect_files=True)
 except exceptions.QCOPBaseError as e:
-    po = e.program_output
-    print(po.stdout)  # or output.pstdout for short
-    print(f"Success: {po.success}")  # False
-    print(po.input_data)  # Input data used to generate the calculation
-    print(po.provenance)  # Provenance of generated calculation
-    print(po.traceback)  # or output.ptraceback for short
+    results = e.results
+    print(results.logs)
+    print(f"Success: {results.success}")  # False
+    print(results.input_data)  # Input data used to generate the calculation
+    print(results.provenance)  # Provenance of generated calculation
+    print(results.traceback)  # or output.ptraceback for short
     raise
 
 else:
     # Check results
-    print(po.stdout)  # or output.pstdout for short
-    print(f"Success: {po.success}")  # True
-    print("output.results: ", po.results)
-    print("output.results.gradient:", po.results.gradient)
-    print(po.input_data)  # Input data used to generate the calculation
-    print(po.provenance)  # Provenance of generated calculation
+    print(results.logs)
+    print(f"Success: {results.success}")  # True
+    print("output.data: ", results.data)
+    print("output.data.gradient:", results.data.gradient)
+    print(results.input_data)  # Input data used to generate the calculation
+    print(results.provenance)  # Provenance of generated calculation
