@@ -1,6 +1,6 @@
 import numpy as np
 import pytest
-from qcio import ProgramInput, Structure
+from qcio import CalcSpec, Structure
 
 from qcop import compute
 from tests.conftest import skipif_program_not_available
@@ -9,7 +9,7 @@ from tests.conftest import skipif_program_not_available
 @pytest.mark.integration
 @skipif_program_not_available("crest")
 def test_crest():
-    input_data = ProgramInput(
+    input_data = CalcSpec(
         structure=Structure(
             symbols=["O", "H", "H"],
             # Integration test depend upon this geometry; do not change
@@ -28,8 +28,8 @@ def test_crest():
 
     prog_output = compute("crest", input_data)
 
-    assert len(prog_output.results.conformers) == 1
-    assert np.isclose(prog_output.results.conformer_energies[0], -0.33568982, atol=1e-4)
+    assert len(prog_output.data.conformers) == 1
+    assert np.isclose(prog_output.data.conformer_energies[0], -0.33568982, atol=1e-4)
     # Solvent getting passed correctly
-    assert "alpb" in prog_output.stdout
-    assert "acetonitrile" in prog_output.stdout
+    assert "alpb" in prog_output.logs
+    assert "acetonitrile" in prog_output.logs
