@@ -1,6 +1,6 @@
 import numpy as np
 import pytest
-from qcio import CalcType, ProgramInput, Structure
+from qcio import CalcSpec, CalcType, Structure
 
 from qcop.main import compute
 from tests.conftest import skipif_program_not_available
@@ -10,7 +10,7 @@ from tests.conftest import skipif_program_not_available
 @skipif_program_not_available("terachem")
 def test_terachem_energy(hydrogen):
     # Modify keywords
-    energy_inp = ProgramInput(
+    energy_inp = CalcSpec(
         structure=hydrogen,
         calctype=CalcType.energy,
         model={"method": "hf", "basis": "sto-3g"},
@@ -20,14 +20,14 @@ def test_terachem_energy(hydrogen):
     output = compute(program, energy_inp)
     assert output.input_data == energy_inp
     assert output.provenance.program == program
-    assert np.isclose(output.results.energy, -1.1167143325, atol=1e-6)
+    assert np.isclose(output.data.energy, -1.1167143325, atol=1e-6)
 
 
 @pytest.mark.integration
 @skipif_program_not_available("terachem")
 def test_terachem_gradient(hydrogen):
     # Modify keywords
-    energy_inp = ProgramInput(
+    energy_inp = CalcSpec(
         structure=hydrogen,
         calctype=CalcType.gradient,
         model={"method": "hf", "basis": "sto-3g"},
@@ -38,9 +38,9 @@ def test_terachem_gradient(hydrogen):
     output = compute(program, energy_inp)
     assert output.input_data == energy_inp
     assert output.provenance.program == program
-    assert np.isclose(output.results.energy, -1.1167143325, atol=1e-6)
+    assert np.isclose(output.data.energy, -1.1167143325, atol=1e-6)
     assert np.allclose(
-        output.results.gradient,
+        output.data.gradient,
         np.array([[0.0, 0.0, -0.02845402], [0.0, 0.0, 0.02845402]]),
         atol=1e-6,
     )
@@ -59,7 +59,7 @@ def test_terachem_hessian():
             [1.147442147965631, -0.49902392855765926, -1.3798296942644486],
         ],
     )
-    energy_inp = ProgramInput(
+    energy_inp = CalcSpec(
         structure=h2o,
         calctype=CalcType.hessian,
         model={"method": "b3lyp", "basis": "6-31g"},
@@ -74,11 +74,11 @@ def test_terachem_hessian():
     assert result.provenance.program == program
 
     # Energy assertion
-    assert np.isclose(result.results.energy, -76.3861099088, atol=1e-6)
+    assert np.isclose(result.data.energy, -76.3861099088, atol=1e-6)
 
     # Gradient assertion
     assert np.allclose(
-        result.results.gradient,
+        result.data.gradient,
         np.array(
             [
                 [-2.69528e-05, -3.88595e-05, 3.06421e-05],
@@ -91,7 +91,7 @@ def test_terachem_hessian():
 
     # Hessian assertion
     assert np.allclose(
-        result.results.hessian,
+        result.data.hessian,
         np.array(
             [
                 [
@@ -203,7 +203,7 @@ def test_terachem_hessian():
 @skipif_program_not_available("terachem")
 def test_terachem_optimization(water):
     # Modify keywords
-    energy_inp = ProgramInput(
+    energy_inp = CalcSpec(
         structure=water,
         calctype=CalcType.gradient,
         model={"method": "hf", "basis": "sto-3g"},
