@@ -1,5 +1,5 @@
 import pytest
-from qcio import CalcType, ProgramOutput
+from qcio import CalcType, Results, SinglePointData
 
 from qcop.adapters import base, registry
 from qcop.exceptions import AdapterInputError, ExternalProgramError
@@ -14,7 +14,7 @@ def test_adapter_subclasses_must_define_program():
             # supported_calctypes defined but program not defined
             supported_calctypes = [CalcType.energy]
 
-            def compute_results(self, *args, **kwargs):
+            def compute_data(self, *args, **kwargs):
                 pass
 
 
@@ -27,7 +27,7 @@ def test_adapter_subclasses_must_define_supported_calctypes():
             # program defined but supported_calctypes not defined
             program = "test"
 
-            def compute_results(self, *args, **kwargs):
+            def compute_data(self, *args, **kwargs):
                 pass
 
 
@@ -53,13 +53,13 @@ def test_adapter_subclasses_defining_program_and_supported_calctypes():
         program = "test"
         supported_calctypes = [CalcType.energy]
 
-        def compute_results(self, *args, **kwargs):
+        def compute_data(self, *args, **kwargs):
             pass
 
     assert registry.get("test") == TestAdapter
 
 
-def test_adapters_raise_error_if_calctype_not_supported(prog_inp):
+def test_adapters_raise_error_if_calctype_not_supported(calcspec):
     """Test that adapters raise an error if the calctype is not supported."""
 
     class TestAdapter(base.ProgramAdapter):
@@ -70,77 +70,74 @@ def test_adapters_raise_error_if_calctype_not_supported(prog_inp):
         def program_version(self, *args, **kwargs) -> str:
             return "v1.0.0"
 
-        def compute_results(self, *args, **kwargs):
+        def compute_data(self, *args, **kwargs):
             pass
 
-    gradient_input = prog_inp("gradient")
+    gradient_input = calcspec("gradient")
     with pytest.raises(AdapterInputError):
         TestAdapter().compute(gradient_input)
 
 
-def test_results_added_to_program_output_object_if_exception_contains_them(
-    prog_inp, mocker, prog_output, test_adapter
+def test_data_added_to_results_object_if_exception_contains_it(
+    calcspec, mocker, results, test_adapter
 ):
-    """Test that results are added to the ProgramOutput object if the exception
+    """Test that results are added to the Results object if the exception
     contains them."""
     test_adapter = registry["test"]()
 
     def raise_error(*args, **kwargs):
         raise ExternalProgramError(
-            program="terachem", stdout="some stdout", results=prog_output.results
+            program="terachem", logs="some stdout", data=results.data
         )
 
     mocker.patch.object(
         test_adapter,
-        "compute_results",
+        "compute_data",
         side_effect=raise_error,
     )
-    energy_input = prog_inp("energy")
+    energy_input = calcspec("energy")
 
     # Check that the exception object contains the results
     with pytest.raises(ExternalProgramError) as excinfo:
         test_adapter.compute(energy_input, raise_exc=True)
-    assert excinfo.value.results == prog_output.results
+    assert excinfo.value.data == results.data
 
-    # If no raise_exc=False, the results are added to the ProgramOutput
-    prog_failure = test_adapter.compute(energy_input, raise_exc=False)
-    assert isinstance(prog_failure, ProgramOutput)
-    assert prog_failure.results == prog_output.results
+    # If no raise_exc=False, the results are added to the Results
+    computed_results = test_adapter.compute(energy_input, raise_exc=False)
+    assert isinstance(computed_results, Results)
+    assert computed_results.data == results.data
 
 
-def test_program_output_object_added_to_exception(
-    prog_inp, mocker, prog_output, test_adapter
-):
-    """Test that exceptions contain the ProgramOutput object."""
+def test_results_object_added_to_exception(calcspec, mocker, results, test_adapter):
+    """Test that exceptions contain the Results object."""
     test_adapter = registry["test"]()
 
     def raise_error(*args, **kwargs):
         raise ExternalProgramError(
             program="terachem",
-            stdout="some stdout",
-            results=prog_output.results,
+            logs="some stdout",
+            data=results.data,
         )
 
     mocker.patch.object(
         test_adapter,
-        "compute_results",
+        "compute_data",
         side_effect=raise_error,
     )
-    energy_input = prog_inp("energy")
+    energy_input = calcspec("energy")
 
     # Check that the exception object contains the results
     with pytest.raises(ExternalProgramError) as excinfo:
         test_adapter.compute(energy_input, raise_exc=True)
 
-    assert isinstance(excinfo.value.program_output, ProgramOutput)
-    assert excinfo.value.program_output.success is False
+    assert isinstance(excinfo.value.results, Results)
+    assert excinfo.value.results.success is False
     # NOTE: CHECK WITH BIGCHEM
-    # assert isinstance(excinfo.value.args[-1], ProgramOutput)
-    
+    assert isinstance(excinfo.value.args[-1], SinglePointData)
 
 
 def test_stdout_collected_with_failed_execution(
-    prog_inp, mocker, prog_output, test_adapter
+    calcspec, mocker, results, test_adapter
 ):
     """Test that stdout is collected even if the execution fails."""
     test_adapter = registry["test"]()
@@ -148,27 +145,27 @@ def test_stdout_collected_with_failed_execution(
     def raise_error(*args, **kwargs):
         raise ExternalProgramError(
             program="terachem",
-            stdout="some stdout",
-            results=prog_output.results,
+            logs="some stdout",
+            data=results.data,
         )
 
     mocker.patch.object(
         test_adapter,
-        "compute_results",
+        "compute_data",
         side_effect=raise_error,
     )
-    energy_input = prog_inp("energy")
+    energy_input = calcspec("energy")
 
     # Check that the exception object contains the results
     with pytest.raises(ExternalProgramError) as excinfo:
         test_adapter.compute(energy_input, raise_exc=True)
 
     # Added to exception
-    assert excinfo.value.stdout == "some stdout"
-    # Added to ProgramOutput
-    assert excinfo.value.program_output.stdout == "some stdout"
+    assert excinfo.value.logs == "some stdout"
+    # Added to Results
+    assert excinfo.value.results.logs == "some stdout"
     # Added to exception
-    assert excinfo.value.results == prog_output.results
+    assert excinfo.value.data == results.data
 
 
 def test_collect_wfn_raises_adapter_input_error_if_not_implemented(test_adapter):
