@@ -20,7 +20,7 @@ def test_compute_raises_program_not_found_if_adapter_but_no_program_in_qcng():
         check_qcng_support("mrchem")  # Assumes mrchem is not installed
 
 
-def test_qcng_fallback_tried_if_adapter_not_in_qcop(mocker, prog_inp):
+def test_qcng_fallback_tried_if_adapter_not_in_qcop(mocker, calcspec):
     # WOW: Just a crazy note. This test takes ~1.2 to execute, this is the startup
     # overhead cost of qcng.compute(). Absolutely crazy! It comes from the very slow
     # get_program() call, which takes >1s to execute.
@@ -35,7 +35,7 @@ def test_qcng_fallback_tried_if_adapter_not_in_qcop(mocker, prog_inp):
         qcng_adapter_cls, "program_version", return_value="fake-version"
     )
 
-    energy_inp = prog_inp("energy")
+    energy_inp = calcspec("energy")
     # Program not in qcop, but in qcn
     # Will raise qcengine.exceptions.ResourceError: Program mrchem is registered with
     # QCEngine, but cannot be found. Test still demonstrates that qcng.compute() is
@@ -46,32 +46,32 @@ def test_qcng_fallback_tried_if_adapter_not_in_qcop(mocker, prog_inp):
     assert compute_spy.call_count == 1
 
 
-def test_qcng_compute_not_called_if_adapter_in_qcop(mocker, prog_inp, test_adapter):
+def test_qcng_compute_not_called_if_adapter_in_qcop(mocker, calcspec, test_adapter):
     qcng_spy = mocker.patch("qcengine.compute")
 
     test_adapter = registry["test"]
     compute_spy = mocker.spy(test_adapter, "compute")
 
-    energy_inp = prog_inp("energy")
+    energy_inp = calcspec("energy")
     compute("test", energy_inp)
 
     assert qcng_spy.call_count == 0  # qcng.compute() not called
     assert compute_spy.call_count == 1  # adaptor.compute() called
 
 
-def test_qcng_exception_wrapping_raise_exc_true(mocker, prog_inp):
+def test_qcng_exception_wrapping_raise_exc_true(mocker, calcspec):
     # So system check passes for harness and program installation
     mocker.patch("qcop.utils.check_qcng_support")
     # qcng_spy = mocker.patch("qcengine.compute")
     # qcng_spy.side_effect = QCEngineException("QCEngine Failed!")
 
-    energy_inp = prog_inp("energy")
+    energy_inp = calcspec("energy")
     # Program not in qcop, but in qcng
     with pytest.raises(ExternalProgramError):
         compute("mrchem", energy_inp, raise_exc=True)
 
 
-def test_qcng_exception_wrapping_raise_exc_false(mocker, prog_inp):
+def test_qcng_exception_wrapping_raise_exc_false(mocker, calcspec):
     # So system check passes for harness and program installation
     mocker.patch("qcop.utils.check_qcng_support")
     qcng_adapter = registry["qcengine"]
@@ -79,7 +79,7 @@ def test_qcng_exception_wrapping_raise_exc_false(mocker, prog_inp):
     # Mock the "program_version" method and set its return value
     mocker.patch.object(qcng_adapter, "program_version", return_value="fake-version")
 
-    energy_inp = prog_inp("energy")
+    energy_inp = calcspec("energy")
     # Program not in qcop, but in qcng
     output = compute("mrchem", energy_inp, raise_exc=False)
     assert output.success is False
